isl_multi_templ.c: extract out isl_multi_from_base_templ.c
This makes the function isl_multi_*_from_* opt-in rather than opt-out.
