+++ Makefile.am
@@ -444,6 +444,7 @@ EXTRA_DIST = \
 	isl_multi_dims.c \
 	isl_multi_domain_templ.c \
 	isl_multi_floor.c \
+	isl_multi_from_base_templ.c \
 	isl_multi_gist.c \
 	isl_multi_hash.c \
 	isl_multi_intersect.c \
