+++ isl_aff.c
@@ -3898,6 +3898,7 @@ error:
 #include <isl_multi_cmp.c>
 #include <isl_multi_dims.c>
 #include <isl_multi_floor.c>
+#include <isl_multi_from_base_templ.c>
 #include <isl_multi_gist.c>
 #include <isl_multi_identity_templ.c>
 #include <isl_multi_nan_templ.c>
@@ -6415,6 +6416,7 @@ error:
 #include <isl_multi_coalesce.c>
 #include <isl_multi_domain_templ.c>
 #include <isl_multi_dims.c>
+#include <isl_multi_from_base_templ.c>
 #include <isl_multi_gist.c>
 #include <isl_multi_hash.c>
 #include <isl_multi_identity_templ.c>
@@ -8289,6 +8291,7 @@ error:
 #include <isl_multi_arith_templ.c>
 #include <isl_multi_coalesce.c>
 #include <isl_multi_floor.c>
+#include <isl_multi_from_base_templ.c>
 #include <isl_multi_gist.c>
 #include <isl_multi_align_set.c>
 #include <isl_multi_align_union_set.c>
