+++ isl_val.c
@@ -1661,7 +1661,6 @@ isl_stat isl_val_check_match_domain_space(__isl_keep isl_val *v,
 #undef BASE
 #define BASE val
 
-#define NO_FROM_BASE
 #define NO_MOVE_DIMS
 #include <isl_multi_no_explicit_domain.c>
 #include <isl_multi_templ.c>
